requirements.txt
# Core
python-dotenv>=1.0.0
pydantic>=2.0.0

# Document processing
pymupdf>=1.23.0
python-docx>=1.1.0

# Embeddings — local, free
sentence-transformers>=3.0.0

# Vector store — local, free
chromadb>=0.5.0

# Topic modelling — local, free
bertopic>=0.16.0
scikit-learn>=1.4.0
umap-learn>=0.5.6
hdbscan>=0.8.33

# LLM — Groq free API
groq>=0.9.0

# Speech-to-text — local Whisper, free
openai-whisper>=20231117
soundfile>=0.12.1

# Text-to-speech — free (needs internet, no key)
gTTS>=2.5.0

# Voice input in Streamlit
streamlit-mic-recorder>=0.0.8

# App
streamlit>=1.35.0
plotly>=5.22.0
pandas>=2.2.0
numpy>=1.26.0

# Evaluation
rouge-score>=0.1.2
datasets>=2.19.0
